backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
django-storages>=1.14
Pillow>=10.0
psycopg2-binary>=2.9
celery>=5.4
redis>=5.0
python-decouple>=3.8
gunicorn>=22.0
qrcode[pil]>=7.4
